# docs/api/openapi-backend.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mos-Transport Dispatcher Backend",
  "description": "Приём потока телеметрии NDTP (TCP), сопоставление с эталонным расписанием, расчёт производных признаков, оркестрация прогнозов ML-сервиса (горизонт 10–15 мин), риск/инциденты для диспетчера.\n\nWebSocket `/ws` — поток снимков состояния для дашборда (≈1 раз в секунду).\n\nРасширения: ETA на ближайшие остановки, what-if моделирование, таймлайн, реестр реальных маршрутов (OSM/GTFS), экспорт GTFS-Realtime/GTFS/GeoJSON для картографических сервисов.",
  "version": "2.0.0"
 },
 "paths": {
...
  "/api/ndtp/parse": {
   "post": {
    "tags": [
     "ndtp"
    ],
    "summary": "Разобрать NDTP-кадр(ы) из hex (отладка парсера)",
    "description": "Разбирает NDTP-кадры тем же декодером, что и TCP-сервер. Пример — реальный кадр эмулятора.\n\nArgs:\n    hex_data: Кадры в hex; пробелы и переводы строк игнорируются.\n\nReturns:\n    dict: ``frames`` (заголовки, ячейки, навигация), ``errors``, ``last_error``.\n\nRaises:\n    HTTPException: 400 — строка не является hex.",
    "operationId": "parse_ndtp_api_ndtp_parse_post",
    "requestBody": {
     "content": {
      "text/plain": {
       "schema": {
        "type": "string",
        "title": "Hex Data",
        "description": "Один или несколько NDTP-кадров в hex (пробелы допускаются)",
        "examples": [
         "7e7e26000000c15c0200cc110000000100650001000200000000001ba1b76a0d2b5f1609473821e0842d0030003d01060df0000e02"
        ]
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/vehicles/{vehicle_id}": {
   "get": {
    "tags": [
     "dispatcher"
    ],
    "summary": "Детали ТС (unit_id или tr_id): признаки модели, трейл прогнозов, факт прибытий",
    "description": "Карточка ТС: прогноз, SHAP-объяснение, признаки модели, ETA, история прогнозов и прибытий.\n\nArgs:\n    vehicle_id: unit_id терминала или tr_id ТС.\n\nRaises:\n    HTTPException: 404 — ТС не найдено.",
    "operationId": "vehicle_api_vehicles__vehicle_id__get",
    "parameters": [
     {
      "name": "vehicle_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vehicle Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/incidents/{incident_id}/ack": {
   "post": {
    "tags": [
     "dispatcher"
    ],
    "summary": "Принять инцидент в работу",
    "description": "Отмечает инцидент принятым диспетчером.\n\nArgs:\n    incident_id: ID инцидента из ``/api/incidents``.\n\nRaises:\n    HTTPException: 404 — инцидент не найден или уже закрыт.",
    "operationId": "ack_api_incidents__incident_id__ack_post",
    "parameters": [
     {
      "name": "incident_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Incident Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}